_before/VQA_cslee/requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.8
async-timeout==4.0.2
attrs==19.3.0
Automat==0.8.0
bitarray==2.4.0
black==22.1.0
blinker==1.4
cachetools==5.0.0
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
cloud-init==21.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cycler==0.11.0
Cython==0.29.28
datasets==2.0.0
dbus-python==1.2.16
dill==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
einops==0.4.1
entrypoints==0.3
filelock==3.6.0
flake8==4.0.1
fonttools==4.31.2
frozenlist==1.3.0
fsspec==2022.2.0
ftfy==6.0.3
future==0.18.2
google-auth==2.6.2
google-auth-oauthlib==0.4.6
grpcio==1.45.0
httplib2==0.14.0
huggingface-hub==0.4.0
hydra-core==1.0.7
hyperlink==19.0.0
idna==2.8
importlib-metadata==4.11.3
importlib-resources==5.4.0
incremental==16.10.1
iniconfig==1.1.1
Jinja2==2.10.1
joblib==1.1.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.4.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Markdown==3.3.6
MarkupSafe==1.1.0
matplotlib==3.5.1
mccabe==0.6.1
more-itertools==4.2.0
multidict==6.0.2
multiprocess==0.70.12.2
mypy-extensions==0.4.3
netifaces==0.10.4
nltk==3.7
nodeenv==1.6.0
numpy==1.22.3
oauthlib==3.1.0
omegaconf==2.0.6
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.1
pathspec==0.9.0
pexpect==4.6.0
Pillow==9.0.1
platformdirs==2.5.1
pluggy==1.0.0
portalocker==2.4.0
protobuf==3.19.4
py==1.11.0
pyarrow==7.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycocoevalcap==1.2
pycocotools==2.0.4
pycodestyle==2.8.0
pycparser==2.21
pyDeprecate==0.3.1
pyflakes==2.4.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.7
pyright==1.1.231
pyrsistent==0.15.5
pyserial==3.4
pytest==7.1.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytorch-lightning==1.5.10
pytz==2022.1
PyYAML==5.3.1
regex==2022.3.15
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
responses==0.18.0
rouge-score==0.0.4
rsa==4.8
sacrebleu==2.0.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.2
ssh-import-id==5.10
systemd-python==234
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.4.1
timm==0.5.4
tomli==2.0.1
torch==1.11.0
torchaudio==0.11.0+cu113
torchmetrics==0.7.3
torchvision==0.12.0
tqdm==4.63.1
Twisted==18.9.0
typing-extensions==4.1.1
ubuntu-advantage-tools==27.6
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.9
wadllib==1.3.3
wcwidth==0.2.5
Werkzeug==2.0.3
xxhash==3.0.0
yarl==1.7.2
zipp==3.7.0
zope.interface==4.7.1
